+++ test/radiosity/element3.test.js
@@ -1,8 +1,8 @@
-import Element3 from '../radiosity/element3.js';
-import Point3 from '../radiosity/point3.js';
-import Vertex3 from '../radiosity/vertex3.js';
-import Vector3 from '../radiosity/vector3.js';
-import Spectra from '../radiosity/spectra.js';
+import Element3 from '../../radiosity/element3.js';
+import Point3 from '../../radiosity/point3.js';
+import Vertex3 from '../../radiosity/vertex3.js';
+import Vector3 from '../../radiosity/vector3.js';
+import Spectra from '../../radiosity/spectra.js';
 
 // ^ y
 // |
