+++ test/radiosity/environment.test.js
@@ -1,4 +1,4 @@
-import Environment from '../radiosity/environment.js';
+import Environment from '../../radiosity/environment.js';
 
 test('todo', () => {
   expect(Environment).toBeDefined();
