+++ test/radiosity/instance.test.js
@@ -1,4 +1,4 @@
-import Instance from '../radiosity/instance.js';
+import Instance from '../../radiosity/instance.js';
 
 test('todo', () => {
   expect(Instance).toBeDefined();
