+++ test/radiosity/surface3.test.js
@@ -1,4 +1,4 @@
-import Surface3 from '../radiosity/surface3.js';
+import Surface3 from '../../radiosity/surface3.js';
 
 test('todo', () => {
   expect(Surface3).toBeDefined();
